requirements.txt

viam-sdk==0.47.0

typing-extensions
bson
json
